backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# MongoDB
pymongo>=4.5.0
motor>=3.3.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Langchain for Gemini integration
langchain>=0.0.267
langchain-google-genai>=0.0.3
google-generativeai>=0.3.0

# Utilities
python-multipart>=0.0.6
httpx>=0.24.1 
pyjwt